moving activitySchema to validate folder
